requirements-dev.txt
mypy
psycopg
pylint
pytest-cov
